Déplace opponent et caracteristic dans character
