+++ core/src/main/java/io/github/algorys/agshmne/character/opponent/OpponentDifficultyLevel.java
@@ -1,4 +1,4 @@
-package io.github.algorys.agshmne.opponent;
+package io.github.algorys.agshmne.character.opponent;
 
 public enum OpponentDifficultyLevel {
 	low, medium, hard, mythic
